refactor: fix server data and co-locate menu saga, selector and reducers
relates to #2
