+++ server/data/item-categories.ts
@@ -1,20 +1,23 @@
-import ItemCategory from '../../src/shared/models/ItemCategory';
+import MenuCategory from '../../src/menu/models/MenuCategory';
 
-const itemCategories: ItemCategory[] = [
+const itemCategories: MenuCategory[] = [
   {
     id: '4afb73bb-1465-4c4b-b421-d58af3b55757',
     name: 'Meat & first dishes',
     description: 'Slow cook meals and fresh fishes',
+    order: 1
   },
   {
     id: 'fcf5bbcb-0ca1-433e-addd-5efe1d0f61a0',
     name: 'Seasonal and burgers',
     description: 'Genuine and fresh ingredients',
+    order: 2
   },
   {
     id: '9db2ee06-3ebc-494f-820f-afcced494adc',
     name: 'Desserts & creams',
     description: 'Hand made ice creams and desserts',
+    order: 3
   },
 ];
